+++ src/main/java/com/cobber/fta/plugins/LogicalTypeIPAddress.java
@@ -79,7 +79,7 @@ public class LogicalTypeIPAddress extends LogicalTypeInfinite {
 
 	@Override
 	public String isValidSet(String dataStreamName, long matchCount, long realSamples, StringFacts stringFacts,
-			Map<String, Integer> cardinality, Map<String, Integer> outliers) {
+			Map<String, Long> cardinality, Map<String, Long> outliers) {
 		return (double) matchCount / realSamples >= getThreshold() / 100.0 ? null : ".+";
 	}
